19.02 1. Problem with filter ( string 78-102 ) 2. the rest seems fine  ( остальное вроде бы нормально )

+++ src/main/java/newDoner/papa/one/components/food/Soda.java
@@ -2,14 +2,14 @@ package newDoner.papa.one.components.food;
 
 public class Soda extends OneIngredientFood {
 
-    private final int volume;
+    private final double volume;
 
-    public Soda(String name, int weight, double price, int volume) {
+    public Soda(String name, int weight, double price, double volume) {
         super(name, weight, price);
         this.volume = volume;
     }
 
-    public int getVolume() {
+    public double getVolume() {
         return volume;
     }
